Move all relevant Pangaea constants into own class (as they're shared across Item, ItemLogistics etc)
