+++ src/Item/ItemLogistics.php
@@ -2,6 +2,7 @@
 namespace Pangaea\Item;
 
 use \Pangaea\Item;
+use \Pangaea\Pangaea;
 use \Pangaea\PangaeaException;
 use \Pangaea\RenderableInterface;
 use \Pangaea\Xml;
@@ -65,183 +66,6 @@ class ItemLogistics implements RenderableInterface
     ];
 
     /**
-     * Valid Unit Cost Currencies.
-     *
-     * @const
-     */
-    const UNIT_COST_CURRENCIES = [
-      'AED',
-      'AFN',
-      'ALL',
-      'AMD',
-      'ANG',
-      'AOA',
-      'ARS',
-      'AUD',
-      'AWG',
-      'AZN',
-      'BAM',
-      'BBD',
-      'BDT',
-      'BGN',
-      'BHD',
-      'BIF',
-      'BMD',
-      'BND',
-      'BOB',
-      'BRL',
-      'BSD',
-      'BTN',
-      'BWP',
-      'BYR',
-      'BZD',
-      'CAD',
-      'CDF',
-      'CHF',
-      'CLP',
-      'CNY',
-      'COP',
-      'CRC',
-      'CUP',
-      'CVE',
-      'CZK',
-      'DJF',
-      'DKK',
-      'DOP',
-      'DZD',
-      'EGP',
-      'ERN',
-      'ETB',
-      'EUR',
-      'FJD',
-      'FKP',
-      'GBP',
-      'GEL',
-      'GHS',
-      'GIP',
-      'GMD',
-      'GNF',
-      'GTQ',
-      'GYD',
-      'HKD',
-      'HNL',
-      'HRK',
-      'HTG',
-      'HUF',
-      'IDR',
-      'ILS',
-      'INR',
-      'IQD',
-      'IRR',
-      'ISK',
-      'JMD',
-      'JOD',
-      'JPY',
-      'KES',
-      'KGS',
-      'KHR',
-      'KMF',
-      'KPW',
-      'KRW',
-      'KWD',
-      'KYD',
-      'KZT',
-      'LAK',
-      'LBP',
-      'LKR',
-      'LRD',
-      'LSL',
-      'LTL',
-      'LVL',
-      'LYD',
-      'MAD',
-      'MDL',
-      'MGA',
-      'MKD',
-      'MMK',
-      'MNT',
-      'MOP',
-      'MRO',
-      'MUR',
-      'MVR',
-      'MWK',
-      'MXN',
-      'MYR',
-      'MZN',
-      'NAD',
-      'NGN',
-      'NIO',
-      'NOK',
-      'NPR',
-      'NZD',
-      'OMR',
-      'PAB',
-      'PEN',
-      'PGK',
-      'PHP',
-      'PKR',
-      'PLN',
-      'PYG',
-      'QAR',
-      'RON',
-      'RSD',
-      'RUB',
-      'RUR',
-      'RWF',
-      'SAR',
-      'SBD',
-      'SCR',
-      'SDG',
-      'SEK',
-      'SGD',
-      'SHP',
-      'SLL',
-      'SOS',
-      'SRD',
-      'STD',
-      'SYP',
-      'SZL',
-      'THB',
-      'TJS',
-      'TMT',
-      'TND',
-      'TOP',
-      'TRY',
-      'TTD',
-      'TWD',
-      'TZS',
-      'UAH',
-      'UGX',
-      'USD',
-      'UYU',
-      'UZS',
-      'VEF',
-      'VND',
-      'VUV',
-      'WST',
-      'XAF',
-      'XAG',
-      'XAU',
-      'XBA',
-      'XBB',
-      'XBC',
-      'XBD',
-      'XCD',
-      'XDR',
-      'XFU',
-      'XOF',
-      'XPD',
-      'XPF',
-      'XPT',
-      'XTS',
-      'XXX',
-      'YER',
-      'ZAR',
-      'ZMK',
-      'ZWL',
-    ];
-
-    /**
      * Set the unit cost.
      *
      * @param float $amount
@@ -255,7 +79,7 @@ class ItemLogistics implements RenderableInterface
             throw new PangaeaException('Unit cost currency cannot be blank');
         }
 
-        if (! in_array($currency, static::UNIT_COST_CURRENCIES)) {
+        if (! in_array($currency, Pangaea::UNIT_COST_CURRENCIES)) {
             throw new PangaeaException(sprintf('Invalid unit cost currency "%s"', $currency));
         }
 
@@ -310,7 +134,7 @@ class ItemLogistics implements RenderableInterface
             throw new PangaeaException('Unit cost currency cannot be blank');
         }
 
-        if (! in_array($unit, Item::UNITS_MEASUREMENT)) {
+        if (! in_array($unit, Pangaea::UNITS_MEASUREMENT)) {
             throw new PangaeaException(sprintf('Invalid unit of measurement "%s"', $unit));
         }
